LPS-75792 Move MonitoringWebKeys into internal package
