+++ modules/apps/foundation/portal/portal-monitoring/src/main/java/com/liferay/portal/monitoring/internal/constants/MonitoringWebKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.monitoring.constants;
+package com.liferay.portal.monitoring.internal.constants;
 
 /**
  * @author Michael C. Han
